requirements.txt
dotenv
Flask
langchain
PyPDF2
huggingface_hub
gradio
openai
weaviate-client
gunicorn